# Avoid surrogates when generating `char` using Standard distribution