--- a/src/distributions/other.rs
+++ b/src/distributions/other.rs
@@ -50,7 +50,10 @@ impl Distribution<char> for Standard {
         // reserved for surrogates. This is the size of that gap.
         const GAP_SIZE: u32 = 0xDFFF - 0xD800 + 1;
 
+        // Uniform::new(0, 0x11_0000 - GAP_SIZE) can also be used but it
+        // seemed slower.
         let range = Uniform::new(GAP_SIZE, 0x11_0000);
+
         let mut n = range.sample(rng);
         if n <= 0xDFFF {
             n -= GAP_SIZE;