This probably isn't performance critical but I thought it seemed wasteful to have a loop (that could theoretically go on forever) just to generate a char. The new version also seems faster when I benchmarked it.
````
test misc_gen_chars_old       ... bench:       1,031 ns/iter (+/- 177) = 496 MB/s
test misc_gen_chars_new       ... bench:         920 ns/iter (+/- 978) = 556 MB/s
````