@dhardy 2018-06-20
Since the old version only rejected 0.2% of samples, that doesn't explain the performance increase. Simpler logic (no branching other than the subtraction, which might not even be implemented as a branch) may explain the improvement.

Looks good anyway! @pitdicker?