@dhardy 2018-06-20
Might be faster if you sample from 0 to `0x11_0000 - GAP_SIZE` (depending on optimiser).