@Pazzaz 2018-06-20
Seems to give the same performance.
````
test misc_gen_chars       ... bench:         926 ns/iter (+/- 463) = 552 MB/s
````