@pitdicker 2018-06-20
Good job!

I think the `distr_standard_codepoint` is a bit more reliable than the one you added, or at leasts easier to compare with the benchmarks that measure the performance of only the PRNG, and other distributions. Can you remove the benchmarks?

Benchmark before:
```
test distr_standard_codepoint    ... bench:       2,137 ns/iter (+/- 5) = 1871 MB/s
```

With this PR:
```
test distr_standard_codepoint    ... bench:       1,943 ns/iter (+/- 51) = 2058 MB/s
```

With `Uniform::new(0, 0x11_0000 - GAP_SIZE)` (and the corresponding changes):
```
test distr_standard_codepoint    ... bench:       2,020 ns/iter (+/- 16) = 1980 MB/s
```

Not sure why the different range is a bit slower. It has to do one addition less in the range code. On the other hand the check and compensation for characters in the gap and above (`n > 0xD800`) will be true much more often.

Can you add a comment that this is investigated, and a range with `(GAP_SIZE, 0x11_0000)` seems faster?