@pitdicker 2018-06-20
Somewhat funny: On Reddit there is a there is a [unhappy discussion](https://www.reddit.com/r/rust/comments/8s7gei/unsafe_rust_in_actixweb_other_libraries/) on the amount of unsafe code (and questionable use) in actix-web, yet we start adding more unsafe code :smile:.