@dhardy 2018-06-20
Hmm. My take from a quick glance at that discussion is (a) that `unsafe` is sometimes required and sometimes merely used for performance (as here), and (b) that one must be careful to make `unsafe` uses easy to review and not mask real issues.

In this case the code is quite easy to understand, so not a big issue I think.