@pitdicker 2018-06-20
I agree, and am all for it in this case.