@Pazzaz 2018-06-20
> Can you remove the new benchmark?

Sorry, didn't see the existing one. I've removed it.
> Can you add a comment that this is investigated.

Done