@sicking 2018-06-20
One way to reduce the concern about unsafe code would be to add a `debug_assert!` which checks that char::from_u32(n) would have returned the same result.