@dhardy 2018-06-21
> One way to reduce the concern about unsafe code would be to add a debug_assert! which checks that char::from_u32(n) would have returned the same result.

I was thinking `char::from_u32(n).unwrap_unchecked()` where the latter method does a debug assert. But this requires a new `Option` method!