@TheIronBorn 2018-06-23
Debug assertions aren't run with `cargo test --release`, `#[cfg(any(test, debug_assertions))]` would be better