@sicking 2018-06-23
Yeah. A `unwrap_unchecked()` function isn't really implementable in the language as it currently stands, even with unsafe code. There's no way to unwrap an enum without using a full `match` like `unwrap()` does. I guess what you *could* do is use raw pointers and rely on the binary encoding of the `Option`, but that's shaky enough that I don't think it'll happen.