@dhardy 2018-06-23
I wonder if it would be possible with some ugly trick like transmuting to `enum UncheckedSomeOption<T> { Some(T), None(!) }`. But we're way off topic.

I guess adding `debug_assert!(char::from_u32(n).is_some());` would be reasonable here.